Moved stocks related code to a new blueprint and registed it with the app
