+++ app/stocks/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('stocks', __name__)
+
+from app.stocks import routes
