+++ app/stocks/plot.py
