+++ app/templates/stocks/stock.html
@@ -61,12 +61,12 @@
             <td style="border: 0px;">
                 <br>
                 {% if current_user.is_watching(stock) %}
-                    <form action="{{ url_for('main.unwatch', symbol=stock.symbol) }}" method="POST">
+                    <form action="{{ url_for('stocks.unwatch', symbol=stock.symbol) }}" method="POST">
                         {{ form.hidden_tag() }}
                         {{ form.submit(value='Unwatch')}}
                     </form>
                 {% else %}
-                    <form action="{{ url_for('main.watch', symbol=stock.symbol) }}" method="POST">
+                    <form action="{{ url_for('stocks.watch', symbol=stock.symbol) }}" method="POST">
                         {{ form.hidden_tag() }}
                         {{ form.submit(value='Watch')}}
                     </form>
